+++ src/ESPressio_DiagnosticMonitor.hpp
@@ -1,0 +1,2 @@
+#pragma once
+#include "diagnostics/ESPressio_DiagnosticMonitor.hpp"