+++ src/ESPressio_SerialDiagnostics.hpp
@@ -1,0 +1,3 @@
+#pragma once
+#include "ESPressio_Logging.hpp"
+#include "ESPressio_DiagnosticMonitor.hpp"